added rv32_err_t for error manangement inc directory revamped
