+++ README.md
@@ -88,11 +88,11 @@ $ make run ARGS=build/add.elf ASM=add
 #### Output
 ```sh
 $ ./build/slash_sim build/add.elf add
-Slash-Sim Startup 
+Slash-Sim Startup
 RAM Init done
 ELF found. Reading build/add.elf
 ==================================================
-  Section Name   |  Offset  |  Size   | Address 
+  Section Name   |  Offset  |  Size   | Address
 ==================================================
 .text            | 0x001000 | 0x0001c | 0x00001000
 .data            | 0x002000 | 0x00011 | 0x00002000
@@ -136,7 +136,7 @@ $ hexdump out/add/ram_dump.bin
 
 #### CPU REG Dump
 ```sh
-grep -v "0x00000000" out/add/reg_dump.txt
+$ grep -v "0x00000000" out/add/reg_dump.txt
   sp: 0x00003000
   t0: 0x00000013
   t1: 0x00000023
@@ -145,17 +145,4 @@ grep -v "0x00000000" out/add/reg_dump.txt
   pc: 0x0000101c
 
 ```
-
-### Add an instruction to the build
-
-To add an instruction into the framework we need to do the following:
-1. `opcodes.h`: Add an entry into `rv32_opcode_t` if adding a new opcode.
-    * The decode string shall follow formated string consisting of `$rd`, `$rs1`, `$rs2`, `$i` only.
-    * All immidiate value formats of any instruction structure shall be represented using `$i`.
-2. `opcode_tree.c`: Add entry to `opcode_reg_list` using following macros:
-    * `RV32_OPCODE_NO_SUB`    : API to add an opcode entry to the list with no sub functionality to the opcode.
-    * `RV32_OPCODE_WITH_SUB` : API to add an opcode entry to the list with `n` sub functionality to the opcode.
-    * `RV32_SUB_FUNC_ENTRY` : API to add sub functions in an array.
-3. `rv32_execute.c` : Add callback entry for the instruction execution.
-
