+++ inc/common/elf_reader.h
