+++ inc/common/platmem.h
