+++ src/common/elf_reader.c
